+++ template-parts/post-listing.php
@@ -1,14 +1,16 @@
 <div id="post-listing" class="post-listing">
-    <div class="grid-container">
-        <?php if (have_posts()) : while (have_posts()) : the_post(); ?>
-            <div id="post-<?php the_ID(); ?>" <?php post_class('post ' . (is_single() ? 'post--full' : 'post--brief')); ?>>
+    <?php if (have_posts()) : while (have_posts()) : the_post(); ?>
+        <div id="post-<?php the_ID(); ?>" <?php post_class('post--brief' . (is_sticky() ? ' post--sticky' : '')); ?>>
+            <div class="grid-container">
                 <?php get_template_part('/template-parts/post-format-' . (get_post_format() ?: 'standard')); ?>
-            </div><!-- .post -->
-        <?php endwhile; else : ?>
+            </div><!-- .grid-container -->
+        </div><!-- .post -->
+    <?php endwhile; else : ?>
+        <div class="grid-container">
             <div class="post-listing__no-result-message">
                 <?php _e('No posts were found for your criteria.', 'zenwriter'); ?>
             </div><!-- .post-listing__no-result-message -->
-        <?php endif; ?>
-    </div><!-- .grid-container -->
+        </div><!-- .grid-container -->
+    <?php endif; ?>
 </div><!-- .post-listing -->
 <?php zenwriter_numeric_pagination(); ?>