+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ---
 
 ## Unreleased
+### Added
+- Enso symbol on login page and admin bar
+
 ### Changed
 - Page no longer scrolls when the navigation is open
 - Post author layout