+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enso symbol on login page and admin bar
 
 ### Changed
+- Better styling for buttons
 - Better markup for post meta, tags and author
 - Better typography styling for bold and italic
 - Refactored functions