+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enso symbol on login page and admin bar
 
 ### Changed
+- Improved post listing
 - Better navigation styling
 - Better hover responsiveness
 - Reduced calc() usage