+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## Unreleased
+## [1.2.0] - 2019-09-22
 
 ### Added
 - Option to show post formats in Zenwriter way
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-[Unreleased]: https://github.com/ramonkcom/zenwriter/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/ramonkcom/zenwriter/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/ramonkcom/zenwriter/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/ramonkcom/zenwriter/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/ramonkcom/zenwriter/releases/tag/v1.0.0